partial_ldsc - main function to estimate partial genetic correlations

Estimates unadjusted and partial genetic correlations (and heritabilities, on observed scale), and tests for the difference between the two, for all pairs of conditions.

Usage

partial_ldsc(
  conditions,
  confounder,
  condition.names = NULL,
  confounder.name = NULL,
  ld,
  n.blocks = 200,
  log.name = NULL
)

Arguments

conditions

The path to the files containing munged the GWAS summary statistics for the conditions (character)

confounder

The path to the file containing munged the GWAS summary statistics for the confounder (character)

condition.names

The names of the conditions, default=NULL (character)

confounder.name

The name of the confounder, default=NULL (character)

ld

The path to the folder in which the LD scores used in the analysis are located. Expects LD scores formated as required by the original LD score regression software. (character)

n.blocks

The number of blocks used for block-jackknife, default=200 (numeric)

log.name

The name of the log file, default=NULL (character)

Details

conditions and confounder are required arguments. These input files should have been pre-processed using the munge() function from GenomicSEM.